pyproject.toml
[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "my-project"
version = "1.0.0"
description = "A modern Python project"
authors = ["You <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "my_package", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31"
pydantic = "^2.1"
python-dotenv = "^1.1.0"
openai = "^1.76.2"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
black = "^24.1"
flake8 = "^6.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"
pytest-cov = "^5.0.0"
tox = "^4.15.0"
selenium = "^4.32.0"
